requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
requests
python-dotenv
psycopg2-binary